Add files via upload
Here I am showing you guys what's in my home lab, the virtual machines I have, and the settings that the virtual machines are on. For starters I am using VirtualBox as the general application to hold all of the VM's, another option you have is VMWare instead. I recommend that you do your research on them to see what you prefer. I have all of the VM's on my NAT Network which basically means they aren't exposed to the internet, instead they're on a private network where they can all communicate with each other. I have a kali, ubuntu, metasploitable3, and an opnsense machine. All of the other VM's that are shown are for CTF's in the future that I'm going to delete after I complete them. Just in case you aren't sure as to what metasploitable3 is, it's a machine that has many vulnerable services on it for practice. I mainly use the kali machine for everything and sometimes I use the ubuntu machine. I'm working on getting a windows machine so that I could practice on a windows OS. OPNSense has been my most recent addition to my home lab, OPNSense is a firewall system used for blue team purposes such as blocking ports/ traffic and things of that nature. The next screenshots that you'll see is me running the "ip a" command to see what the ip's are for each of the VM's so that I could see if I could communicate with the other VM's on my NAT Network. I began to ping each VM from each other to make sure that I could reach them. Lastly, I ran a nmap scan from my kali VM against the metasploitable3 and ubuntu machines to see what ports were on them, I didn't want to go into depth as this is just the introduction. #BrickByBrick #UnderdogToTopDog